requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
jsonschema
pydantic
markdown
ollama

